+++ src/DecayRates.cpp
@@ -48,7 +48,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 
 	vector<RadialWF> Orbitals(orbitals.size());
 
-	for (int i = 0; i < orbitals.size(); i++)
+	for (int i = 0; i < static_cast<int>(orbitals.size()); i++)
 	{
 		Orbitals[i].set_N(orbitals[i].N());
 		Orbitals[i].set_L(orbitals[i].L());        // Resets occupancy
@@ -74,7 +74,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 //	Interpolate orbitals on the new grid
 	Interpolation W(6);
 
-	for (int i = 0; i < Orbitals.size(); i++) {
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++) {
 		if (input.Exited_Pot_Model() != "V_N-1" && orbitals[i].occupancy() == 0) continue;
 		if (orbitals[i].F[0] != 0) W.RecalcWF(orbitals[i], lattice, Orbitals[i], Lattice);
 		Infinity = lattice.R(orbitals[i].pract_infinity());
@@ -102,7 +102,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 	double ME = 0;
 
 // main loop over all the possible transitions
-	for (int i = 0; i < Orbitals.size(); i++)
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 	{
 		if (Orbitals[i].occupancy() > 0)
 		{
@@ -110,7 +110,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 
 			density.clear();
 			density.resize(infinity + 1);
-			for (int s = 0; s < density.size(); s++) density[s] = pow(Orbitals[i].F[s], 2);
+			for (int s = 0; s < static_cast<int>(density.size()); s++) density[s] = pow(Orbitals[i].F[s], 2);
 			ME = I.Integrate(&density, 0, infinity);
 
 			j = Orbitals[i].occupancy();
@@ -136,7 +136,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 						if (IntegrateContinuum(Lattice, U, Orbitals, &Continuum, i) < 0) {
 						log << "====================================================================" << endl;
 							log << "Continuum didn't converge: " << endl;
-							for (int i = 0; i < orbitals.size(); i++)
+							for (int i = 0; i < static_cast<int>(orbitals.size()); i++)
 							{
 								log << i + 1 << ") n = " << orbitals[i].N() << " l = " << orbitals[i].L()
 									<< " Energy = " << orbitals[i].Energy << " Occup = " << orbitals[i].occupancy() << endl;
@@ -147,7 +147,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 						density.clear();
 						density.resize(infinity + 1);
 						if (input.Gauge() == "length") {
-								for (int s = 0; s < density.size(); s++)	{
+								for (int s = 0; s < static_cast<int>(density.size()); s++)	{
 								density[s] = Lattice.R(s)*Continuum.F[s] * Orbitals[i].F[s];
 							}
 							ME = I.Integrate(&density, 0, infinity);
@@ -155,7 +155,7 @@ vector<photo> DecayRates::Photo_Ion(double omega, ofstream & log)
 						else {
 						// Velocity gauge.
 							double ang_coeff =  0.5*(Orbitals[i].L() - Continuum.L())*(Orbitals[i].L() + Continuum.L() + 1);
-								for (int s = 0; s < density.size(); s++)	{
+								for (int s = 0; s < static_cast<int>(density.size()); s++)	{
 								density[s] = Continuum.F[s] *(Orbitals[i].G[s] + ang_coeff * Orbitals[i].F[s]/Lattice.R(s)) ;
 							}
 							ME = I.Integrate(&density, 0, infinity)/omega;
@@ -188,12 +188,12 @@ vector<fluor> DecayRates::Fluor()
 
 	Adams I(lattice, 10);
 
-	for (int i = 0; i < orbitals.size(); i++)
+	for (int i = 0; i < static_cast<int>(orbitals.size()); i++)
 	{
 		if (orbitals[i].occupancy() < 4 * orbitals[i].L() + 2)//check if there is a hole
 		{
 			N_h = 4 * orbitals[i].L() + 2 - orbitals[i].occupancy();
-			for (int j = i + 1; j < orbitals.size(); j++)
+			for (int j = i + 1; j < static_cast<int>(orbitals.size()); j++)
 			{
 				if (orbitals[j].L() == orbitals[i].L() + 1 || orbitals[j].L() == orbitals[i].L() - 1)//selection rules
 				{
@@ -208,7 +208,7 @@ vector<fluor> DecayRates::Fluor()
 						density.resize(L_max);
 
 						if (input.Gauge() == "length") {
-							for (int s = 0; s < density.size(); s++)	{
+							for (int s = 0; s < static_cast<int>(density.size()); s++)	{
 								density[s] = lattice.R(s) * orbitals[i].F[s] * orbitals[j].F[s];
 							}
 							ME = I.Integrate(&density, 0, density.size()-1);// Electric dipole matrix element
@@ -216,7 +216,7 @@ vector<fluor> DecayRates::Fluor()
 						else {
 						// Velocity gauge.
 							double ang_coeff =  0.5*(orbitals[i].L() - orbitals[j].L())*(orbitals[i].L() + orbitals[j].L() + 1);
-								for (int s = 0; s < density.size(); s++)	{
+								for (int s = 0; s < static_cast<int>(density.size()); s++)	{
 								density[s] = orbitals[j].F[s] *(orbitals[i].G[s] + ang_coeff * orbitals[i].F[s]/lattice.R(s)) ;
 							}
 							ME = I.Integrate(&density, 0, density.size()-1)/(orbitals[j].Energy - orbitals[i].Energy);
@@ -253,7 +253,7 @@ vector<auger> DecayRates::Auger(vector<int> Max_occ, ofstream & log)
 	assert(Max_occ.size() == orbitals.size());
 	//Create new lattice, core and potential suitable for continuum states
 	double k = 0;
-	for (int i = 0; i < orbitals.size(); i++)//find k that corresponds to shortest wavelength
+	for (int i = 0; i < static_cast<int>(orbitals.size()); i++)//find k that corresponds to shortest wavelength
 	{
 		if (orbitals[i].occupancy() < Max_occ[i]) {
 			k = sqrt(-2 * orbitals[i].Energy);
@@ -284,7 +284,7 @@ vector<auger> DecayRates::Auger(vector<int> Max_occ, ofstream & log)
 
 	//check if there are electrons above current orbital that can fill
 	//empty orbital. If above > 1 auger is possible and even hollow orbital should be interpolated
-	for (int i = 0; i < orbitals.size(); i++) {
+	for (int i = 0; i < static_cast<int>(orbitals.size()); i++) {
 		//if (input.Exited_Pot_Model() != "V_N-1" && orbitals[i].occupancy() == 0) continue;
 		if (input.Exited_Pot_Model() == "V_N-1" || i <= allowed) W.RecalcWF(orbitals[i], lattice, Orbitals[i], Lattice);
 		Infinity = lattice.R(orbitals[i].pract_infinity());
@@ -361,7 +361,7 @@ vector<auger> DecayRates::Auger(vector<int> Max_occ, ofstream & log)
 						Continuum.set_L(l_E); //TODO make sure this is fine -S.P.
 						if (IntegrateContinuum(Lattice, U, Orbitals, &Continuum, f) < 0) {
 							log << "Continuum didn't converge: " << endl;
-							for (int i = 0; i < orbitals.size(); i++)
+							for (int i = 0; i < static_cast<int>(orbitals.size()); i++)
 							{
 								log << i + 1 << ") n = " << orbitals[i].N() << " l = " << orbitals[i].L()
 									<< " Energy = " << orbitals[i].Energy << " Occup = " << orbitals[i].occupancy() << endl;
@@ -469,11 +469,11 @@ vector<double> DecayRates::FT_density(double Q_min, double Q_max, int Q_size)
 		}
 
 		if (Q == 0) {
-			for (int i = 0; i < integr.size(); i++) {
+			for (int i = 0; i < static_cast<int>(integr.size()); i++) {
 				integr[i] = Inp_density[i];
 			}
 		} else {
-			for (int i = 0; i < integr.size(); i++) {
+			for (int i = 0; i < static_cast<int>(integr.size()); i++) {
 				tmp = 2*M_PI*Inp_lattice.R(i)*Q;
 				integr[i] = Inp_density[i]*sin(tmp)/tmp;
 			}
@@ -512,7 +512,7 @@ int DecayRates::IntegrateContinuum(Grid &Lattice, Potential &U, vector<RadialWF>
 	int Core_infinity = 0;
 
 	// Try to optimize practical infinity.
-	for (int i = 0; i < Core.size(); i++) {
+	for (int i = 0; i < static_cast<int>(Core.size()); i++) {
 		if (Core[i].pract_infinity() > Core_infinity) Core_infinity = Core[i].pract_infinity();
 	}
